[Aptos Data Client] Split up tests.
