+++ state-sync/aptos-data-client/src/tests/mod.rs
@@ -1,0 +1,10 @@
+// Copyright © Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+mod advertise;
+mod compression;
+mod mock;
+mod peers;
+mod poller;
+mod priority;
+mod utils;
